requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0
scikit-learn>=1.2.0
numpy>=1.24.0
tqdm>=4.64.0
matplotlib>=3.6.0
seaborn>=0.12.0
jupyter>=1.0.0
